Move DefaultVectorUtilProvider to it's own top-level class - since it will likely grow over time
